Bunch of bug fixes and code cleanup
- Actually fixed bars files inside TitleBG not converting
- Added an option to keep the downloaded files
- Fixed and error caused by Tex2 files, where they would convert, but throw out a "File Not Found" error

+++ converter.py
@@ -19,7 +19,7 @@ import bcfconverter as sound
 import argparse
 import bfstpfixer
 import oead
-
+import traceback
 
 
 # Construct an argument parser
@@ -27,19 +27,26 @@ parser = argparse.ArgumentParser(description="Converts mods in BNP format using 
 parser.add_argument("mods", nargs='+')
 args = parser.parse_args()
 
-supp_formats = [".sbfres", ".sbitemico", ".hkcl", ".hkrg", ".bars", ".bfstm", ".bflim", ".sblarc", "instSize"]
+# Supported formats
+supp_formats = [".sbfres", ".sbitemico", ".hkcl", ".bars", ".bfstm", ".bflim", ".sblarc", "instSize"]
+
+def confirm_prompt(question: str) -> bool:
+    # https://gist.github.com/garrettdreyfus/8153571
+    reply = None
+    while reply not in ("", "y", "n"):
+        reply = input(f"{question} (Y/n): ").lower()
+    return (reply in ("", "N"))
 
 def extract_sarc(sarc: oead.Sarc, sarc_path: Path) -> None:
     Path(sarc_path).mkdir(exist_ok=True)
     for file in sarc.get_files():
-        if not Path(f"{sarc_path}{sep}{file.name}").parent.exists():
-            Path(f"{sarc_path}{sep}{file.name}").parent.mkdir(parents=True)
-        Path(f"{sarc_path}{sep}{file.name}").write_bytes(file.data)
+        if not Path(sarc_path / file.name).parent.exists():
+            Path(sarc_path / file.name).parent.mkdir(parents=True)
+        Path(sarc_path / file.name).write_bytes(file.data)
 
 def write_sarc(sarc: oead.Sarc, sarc_path: Path, sarc_file: Path) -> None:
     # Write the modified files back to the sarc
-    new_sarc = oead.SarcWriter().from_sarc(sarc)
-    new_sarc.set_endianness(oead.Endianness.Little)
+    new_sarc = oead.SarcWriter(endian=oead.Endianness.Little)
     for file in sarc_path.rglob("*.*"):
         new_file = f'{file}'.split(f"{sarc_file.name}{sep}")[1]
         new_sarc.files[new_file] = file.read_bytes()
@@ -57,25 +64,30 @@ def convert_fres(sbfres: Path) -> None:
         http_response = urlopen('https://gamebanana.com/dl/485626')
         zipfile = ZipFile(BytesIO(http_response.read()))
         zipfile.extractall(path='BfresPlatformConverter')
+
     if sbfres.suffixes == ['.Tex1', '.sbfres']:
         bfres = util.unyaz_if_needed(sbfres.read_bytes())
         tex2 = Path(f'{sbfres.parent}{sep}{sbfres.name.split(".")[0]}.Tex2.sbfres')
-        tex2b = util.unyaz_if_needed(tex2.read_bytes())
-        sbfres.write_bytes(bfres)
-        tex2.write_bytes(tex2b)
+        if Path(f'{sbfres.parent}{sep}{sbfres.name.split(".")[0]}.Tex2.sbfres').exists():
+            tex2b = util.unyaz_if_needed(tex2.read_bytes())
+            tex2.write_bytes(tex2b)
+            sbfres.write_bytes(bfres)
+
     elif sbfres.suffixes != ['.Tex2', '.sbfres']:
         bfres = util.unyaz_if_needed(sbfres.read_bytes())
         sbfres.write_bytes(bfres)
+
     if system() == "Windows":
         run(["BfresPlatformConverter\BfresPlatformConverter.exe", sbfres])
     else:
         run(['mono', "BfresPlatformConverter/BfresPlatformConverter.exe", sbfres])
 
     if sbfres.suffixes == ['.Tex1', '.sbfres']:
-        c_bfres = oead.yaz0.compress(Path(f'SwitchConverted{sep}{sbfres.stem.rstrip("1")}.sbfres').read_bytes())
-        sbfres.write_bytes(c_bfres)
-        sbfres.rename(Path(f'{sbfres.parent}{sep}{sbfres.stem.rstrip("1")}.sbfres'))
-        tex2.unlink()
+        if tex2.exists():
+            c_bfres = oead.yaz0.compress(Path(f'SwitchConverted{sep}{sbfres.stem.rstrip("1")}.sbfres').read_bytes())
+            sbfres.write_bytes(c_bfres)
+            sbfres.rename(Path(f'{sbfres.parent}{sep}{sbfres.stem.rstrip("1")}.sbfres'))
+            tex2.unlink()
     elif sbfres.suffixes != ['.Tex2', '.sbfres']:
         c_bfres = oead.yaz0.compress(Path(f'SwitchConverted{sep}{sbfres.name}').read_bytes())
         sbfres.write_bytes(c_bfres)
@@ -83,19 +95,24 @@ def convert_fres(sbfres: Path) -> None:
 
 def convert_havok(actorpack: Path) -> None:
     # Convert havok files unsupported by BCML
-    if not Path('HKXConvert').exists():
+    if not Path('HKXConvert').exists() or Path('HKXConvert.exe').exists():
         # Download HKXConvert if it's not already in the system
         print("Downloading HKXConvert...")
-        filename, headers = urlretrieve('https://github.com/krenyy/HKXConvert/releases/download/1.0.1/HKXConvert', filename='HKXConvert')
-        Path('HKXConvert').chmod(0o755)
-    actor_path = Path(f'{actorpack.name}')
+        if system() == "Windows":
+            filename, headers = urlretrieve('https://github.com/krenyy/HKXConvert/releases/download/1.0.1/HKXConvert.exe', filename='HKXConvert.exe')
+            hkx_c = f".{sep}HKXConvert.exe"
+        else:
+            filename, headers = urlretrieve('https://github.com/krenyy/HKXConvert/releases/download/1.0.1/HKXConvert', filename='HKXConvert')
+            hkx_c = f".{sep}HKXConvert"
+        Path(f'{hkx_c}').chmod(0o755)
+    actor_path = Path(actorpack.name)
     actor = oead.Sarc(util.unyaz_if_needed(actorpack.read_bytes()))
     extract_sarc(actor, actor_path)
     hkxs = actor_path.rglob('*.hk*')
     for hkx in hkxs:
-        run([f'.{sep}HKXConvert', 'hkx2json', hkx])
+        run([f'{hkx_c}', 'hkx2json', hkx])
         hkx.unlink()
-        run([f'.{sep}HKXConvert', 'json2hkx', '--nx', f'{splitext(hkx)[0]}.json'])
+        run([f'{hkx_c}', 'json2hkx', '--nx', f'{splitext(hkx)[0]}.json'])
         Path(f'{splitext(hkx)[0]}.json').unlink()
     if hkxs:
         print("HKX files converted. Saving...")
@@ -115,12 +132,14 @@ def convert_bars(bars: Path, mod_root: Path) -> None:
     try:
         o_bars = util.get_game_file(f"Voice{sep}USen{sep}{bars.name}").read_bytes()
     except FileNotFoundError:
-        o_bars = util.get_game_file(f"Sound{sep}Resource{sep}{bars.name}").read_bytes()
-    except FileNotFoundError:
-        bg_path = bars.parent.parent.parent.name
-        pack_path = util.get_game_file(f'Pack{sep}{bg_path}')
-        pack = oead.Sarc(pack_path.read_bytes())
-        o_bars = pack.get_file(f'{bars}'.split(f'{pack_path.name}{sep}')[1]).data
+        try:
+            o_bars = util.get_game_file(f"Sound{sep}Resource{sep}{bars.name}").read_bytes()
+        except FileNotFoundError:
+            bg_path = bars.parent.parent.parent.name
+            pack_path = util.get_game_file(f'Pack{sep}{bg_path}')
+            pack = oead.Sarc(pack_path.read_bytes())
+            o_bars = pack.get_file(f'{bars}'.split(f'{pack_path.name}{sep}')[1]).data
+
     barstool.extract_from_bars(bars, '>')
     try:
         bfstps = Path(bg_path).rglob('*.bfstp')
@@ -142,7 +161,7 @@ def convert_bars(bars: Path, mod_root: Path) -> None:
 
 def convert_bflim(sblarc: Path):
     blarc = oead.Sarc(util.unyaz_if_needed(sblarc.read_bytes()))
-    blarc_path = Path(f'{sblarc.name}')
+    blarc_path = Path(sblarc.name)
     if any("bflim" in i.name for i in blarc.get_files()):
         stock_pack = util.get_game_file(f"Pack{sep}{sblarc.parent.parent.name}")
         try:
@@ -153,37 +172,37 @@ def convert_bflim(sblarc: Path):
             stock_blarc = oead.Sarc(util.unyaz_if_needed(stock_sblarc.data))
         bntx_file = stock_blarc.get_file(f"timg{sep}__Combined.bntx")
         extract_sarc(blarc, blarc_path)
-        Path(f'{blarc_path}{sep}{bntx_file.name}').write_bytes(bntx_file.data)
+        Path(blarc_path / f'{bntx_file.name}').write_bytes(bntx_file.data)
         for bflim in blarc_path.rglob('*.bflim'):
-            bntx.tex_inject(f'{blarc_path}{sep}{bntx_file.name}', bflim, False)
+            bntx.tex_inject(blarc_path / f'{bntx_file.name}', bflim, False)
             Path(bflim).unlink()
         write_sarc(blarc, blarc_path, sblarc)
         shutil.rmtree(blarc_path)
 
 def convert_files(files, mod_path: Path) -> None:
     for file in files:
-        if file.suffix == ".sbfres" or file.suffix == ".sbitemico":
-            # Convert FRES files
-            convert_fres(file)
+        if file.exists() and file.stat().st_size != 0:
+            if file.suffix == ".sbfres" or file.suffix == ".sbitemico":
+                # Convert FRES files
+                convert_fres(file)
 
-        elif file.suffix == ".sbactorpack":
-            # Convert havok files inside actor packs
-            actor = oead.Sarc(util.unyaz_if_needed(file.read_bytes()))
-            actor_path = Path(f'{file.name}')
-            if any("hkcl" in i.name for i in actor.get_files()) or any("hkrg" in i.name for i in actor.get_files()):
-                convert_havok(file)
+            elif file.suffix == ".sbactorpack":
+                # Convert havok files inside actor packs
+                actor = oead.Sarc(util.unyaz_if_needed(file.read_bytes()))
+                actor_path = Path(f'{file.name}')
+                if any("hkcl" in i.name for i in actor.get_files()) or any("hkrg" in i.name for i in actor.get_files()):
+                    convert_havok(file)
 
-        elif file.suffix == ".bars":
-            # Convert bars files
-            convert_bars(file, mod_path)
+            elif file.suffix == ".bars":
+                # Convert bars files
+                convert_bars(file, mod_path)
 
-        elif file.suffix == ".bfstm":
-            # Convert BFSTM files
-            sound.convExtFile(file, "FSTM", '<')
+            elif file.suffix == ".bfstm":
+                # Convert BFSTM files
+                sound.convExtFile(file, "FSTM", '<')
 
-        elif file.suffix == ".pack" or file.suffix == ".sbeventpack":
-            # Convert files inside of pack files
-            if file.name != "AocMainField.pack":
+            elif file.suffix == ".pack" or file.suffix == ".sbeventpack":
+                # Convert files inside of pack files
                 pack = oead.Sarc(util.unyaz_if_needed(file.read_bytes()))
                 pack_path = Path(f'{file.name}')
                 if any(splitext(i.name)[1] in supp_formats for i in pack.get_files()):
@@ -192,33 +211,37 @@ def convert_files(files, mod_path: Path) -> None:
                     write_sarc(pack, pack_path, file)
                     shutil.rmtree(pack_path)
 
-        elif file.suffix == ".sblarc":
-            # Convert bflim files inside of sblarc files
-            convert_bflim(file)
+            elif file.suffix == ".sblarc":
+                # Convert bflim files inside of sblarc files
+                convert_bflim(file)
 
-        elif file.suffix == ".hkcl":
-            convert_havok(file)
-            
-        elif file.name == "actorinfo.yml":
-            # Convert instSizes inside of the actorinfo
-            convert_instS(file)
+            elif file.suffix == ".hkcl":
+                convert_havok(file)
+                
+            elif file.name == "actorinfo.yml":
+                # Convert instSizes inside of the actorinfo
+                convert_instS(file)
 
-def converter(mod) -> None:
-    try:
-        # Set BCML to Switch mode
-        settings = util.get_settings()
-        settings["wiiu"] = False
-        util.save_settings()
+def clean_up():
+    if Path('HKXConvert').exists():
+        Path('HKXConvert').unlink()
+    shutil.rmtree('BfresPlatformConverter', ignore_errors=True)
+    shutil.rmtree('SwitchConverted', ignore_errors=True)
+    shutil.rmtree('WiiUConverted', ignore_errors=True)
 
-        bnp = Path(mod)
-        mod_path = open_mod(bnp)
-        files = mod_path.rglob('*.*')
+def convert(mod) -> None:
+    # Open the mod
+    bnp = Path(mod)
+    mod_path = open_mod(bnp)
+    files = mod_path.rglob('*.*')
+    try: 
 
         # Run the mod through BCML's automatic converter first 
         warnings = convert_mod(mod_path, False, True)
 
         # Convert supported files
         convert_files(files, mod_path)
+            
         out = bnp.with_name(f"{bnp.stem}_switch.bnp")
         if Path(out).exists():
             Path(out).unlink()
@@ -229,39 +252,49 @@ def converter(mod) -> None:
             f'{mod_path / "*"}',
         ]
         run(x_args)
-    except RuntimeError as e:
-        shutil.rmtree(mod_path, ignore_errors=True)
-        if Path('HKXConvert').exists():
-            Path('HKXConvert').unlink()
-        shutil.rmtree('SwitchConverted', ignore_errors=True)
-        shutil.rmtree('WiiUConverted', ignore_errors=True)
-        shutil.rmtree('BfresPlatformConverter', ignore_errors=True)
-        print(e)
 
-    shutil.rmtree(mod_path, ignore_errors=True)
-    if Path('HKXConvert').exists():
-        Path('HKXConvert').unlink()
-    shutil.rmtree('SwitchConverted', ignore_errors=True)
-    shutil.rmtree('WiiUConverted', ignore_errors=True)
-    shutil.rmtree('BfresPlatformConverter', ignore_errors=True)
-    if warnings:
-        with open("error.log", "w") as file:
-            for warning in warnings:
-                # Write all files that couldn't be converted to a file
-                if all(i not in warning for i in supp_formats):
-                    file.write(warning + "\n")
+        # Write BCML's warning to a file
+        if warnings:
+            with open("error.log", "a") as file:
+                for warning in warnings:
+                    # Write BCML's warning to a file
+                    if type(warning) == list:
+                        for error in warning:
+                            if all(i not in error for i in supp_formats):
+                                file.write(error + "\n")
+                        continue
+                    if all(i not in warning for i in supp_formats):
+                        file.write(warning + "\n")
+
+    except Exception:
+        print(traceback.format_exc())
+        clean_up()
+
+    finally:
+        shutil.rmtree(mod_path, ignore_errors=True)
 
 def main() -> None:
     if len(args.mods) == 1: # one argument
         mods = glob(args.mods[0])
     else: # more than one argument
     	mods = args.mods
+
+    downloaded = False
+    if Path('BfresPlatformConverter').exists() and Path('HKXConvert').exists():
+        downloaded = True
+
+    # Set BCML to Switch mode
+    settings = util.get_settings()
+    settings["wiiu"] = False
+    util.save_settings()
     
     for mod in mods:
-        try:
-            converter(mod)
-        except RuntimeError as e:
-            print(e)
+        convert(mod)
+
+    if (Path('BfresPlatformConverter').exists() or Path('HKXConvert').exists()) and not downloaded:
+        reply = confirm_prompt("Would you like to keep the downloaded files?")
+        if not reply:
+            clean_up()
 
 
 if __name__ == "__main__":
